feat: moved ReadFile to common-utils
BREAKING CHANGE: Moved ReadFile to common-utils so it can be used without loading the
full common-config-loader. A few eik project may benefit from that, ie core, sink-gcs.
