+++ packages/config-loader/index.js
@@ -1,11 +1,9 @@
-const ReadFile = require('./src/classes/read-file.js');
 const EikConfig = require('./src/eik-config.js');
 const getDefaults = require('./src/get-defaults.js');
 const configStore = require('./src/config-store.js');
 const localAssets = require('./src/local-assets');
 
 module.exports = {
-    ReadFile,
     EikConfig,
     getDefaults,
     configStore,
